requirements.txt
beautifulsoup4==4.12.3
easyocr==1.7.1
Flask==3.0.3
Flask_SocketIO==5.3.6
numpy==1.26.4
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==10.3.0
pytesseract==0.3.10
Requests==2.31.0
selenium==4.20.0
Werkzeug==3.0.3
